requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile
#
asyncio-mqtt==0.16.2
    # via -r requirements.in
attrs==23.1.0
    # via cmd2
cmd2==1.5.0
    # via kuksa-client
colorama==0.4.6
    # via cmd2
decorator==5.1.1
    # via jsonpath-ng
grpcio==1.58.0
    # via grpcio-tools
grpcio-tools==1.58.0
    # via kuksa-client
jsonpath-ng==1.5.3
    # via kuksa-client
kuksa-client==0.4.0
    # via -r requirements.in
paho-mqtt==1.6.1
    # via asyncio-mqtt
ply==3.11
    # via jsonpath-ng
protobuf==4.24.3
    # via grpcio-tools
pygments==2.16.1
    # via kuksa-client
pyperclip==1.8.2
    # via cmd2
six==1.16.0
    # via jsonpath-ng
typing-extensions==4.7.1
    # via asyncio-mqtt
wcwidth==0.2.6
    # via cmd2
websockets==11.0.3
    # via kuksa-client

# The following packages are considered to be unsafe in a requirements file:
# setuptools
